dom/xul/test/test_bug398289.html

<!DOCTYPE html>
<html style="height: 100%">
<head>
  <title>Test for bug 398289</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/WindowSnapshot.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />
</head>
<body style="height: 100%" onload="setTimeout(onBodyLoad, 0);">
  <iframe id="test" src="398289-resource.xul" width="100%" height="100%">
  </iframe>

  <script class="testbody" type="text/javascript">
    var snap1, snap2;
  
    SimpleTest.waitForExplicitFinish();
  
    function onBodyLoad() {
      window.frames[0].document.getElementById("test").selectedIndex = 0;
      window.frames[0].document.getElementById("test").selectedIndex = 1;
      
      snap1 = snapshotWindow(window);
      
      document.getElementById("test").onload = onFrameLoad;
      window.frames[0].location.reload();
    }
    
    function onFrameLoad() {
      snap2 = snapshotWindow(window);

      var equal, str1, str2;
      [equal, str1, str2] = compareSnapshots(snap1, snap2, true);

      ok(equal, "persistent attribute in tab box broken, expected: "+str1+" got: "+str2);
          
      SimpleTest.finish();
    }
  </script>
</body>
</html>
